requirements.txt
django-filter==1.0.4
Django==1.10.3
djangorestframework==3.6.4
Markdown==2.6.9
pip==9.0.1
pytz==2017.2
setuptools==36.5.0
wheel==0.30.0
